+++ x/liquidstakeibc/types/msgs.go
@@ -351,6 +351,16 @@ func (m *MsgUpdateHostChain) ValidateBasic() error {
 			if err != nil {
 				return fmt.Errorf("unable to unmarshal flags update string")
 			}
+		case KeyRewardParams:
+			var params RewardParams
+			err := json.Unmarshal([]byte(update.Value), &params)
+			if err != nil {
+				return fmt.Errorf("unable to unmarshal reward params update string")
+			}
+
+			if err := sdk.ValidateDenom(params.Denom); err != nil {
+				return fmt.Errorf("invalid rewards denom: %s", err.Error())
+			}
 		default:
 			return fmt.Errorf("invalid or unexpected update key: %s", update.Key)
 		}